Reorganize some components
