+++ src/components/InfectionTable/style.css
